huge package structure refactoring
some chages in default jdbc.properties
