+++ README.md
@@ -1,5 +1,8 @@
 # web-store
 
-Before running the gradle script (when you first open project at your PC) do not forget to
-rename gradle.default.properies to gradle.properies and change glassfishHome path. Also,
-do not forget to change jdbc.properties.
+#### Before running the gradle script (when you first open project at your PC)
+
+Copy **./gradle.default.properies** to **./gradle.properies** and change glassfishHome path.
+
+Also, do not forget to change copy **./src.main.resources/jdbc.default.properties** to
+ **./src.main.resources/jdbc.properties** and chage your DB connection settings
