+++ src/main/java/fedorchuck/com/github/webstore/web/WebConfig.java
@@ -24,6 +24,7 @@ package fedorchuck.com.github.webstore.web;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.ComponentScan;
 import org.springframework.context.annotation.Configuration;
+import org.springframework.http.converter.StringHttpMessageConverter;
 import org.springframework.web.multipart.MultipartResolver;
 import org.springframework.web.multipart.support.StandardServletMultipartResolver;
 import org.springframework.web.servlet.ViewResolver;
@@ -40,12 +41,18 @@ import org.thymeleaf.templateresolver.ServletContextTemplateResolver;
 import org.thymeleaf.templateresolver.TemplateResolver;
 
 import java.io.IOException;
+import java.nio.charset.Charset;
 
 @Configuration
 @EnableWebMvc
 @ComponentScan("fedorchuck.com.github.webstore.web")
 public class WebConfig extends WebMvcConfigurerAdapter {
 
+	// It don't helps
+    @Bean
+    public StringHttpMessageConverter stringHttpMessageConverter() {
+        return new StringHttpMessageConverter(Charset.forName("UTF-8"));
+    }
     @Bean
     public ViewResolver viewResolver(SpringTemplateEngine templateEngine) {
         ThymeleafViewResolver viewResolver = new ThymeleafViewResolver();
