+++ src/main/java/fedorchuck/com/github/webstore/web/controllers/HomeController.java
@@ -19,12 +19,12 @@
  * @license GPL-3.0+ <http://spdx.org/licenses/GPL-3.0+>
  */
 
-package fedorchuck.com.github.webstore.web;
+package fedorchuck.com.github.webstore.web.controllers;
 
-import fedorchuck.com.github.webstore.Category;
-import fedorchuck.com.github.webstore.Commodity;
-import fedorchuck.com.github.webstore.SearchRequest;
-import fedorchuck.com.github.webstore.data.CommodityRepository;
+import fedorchuck.com.github.webstore.domainmodels.Category;
+import fedorchuck.com.github.webstore.domainmodels.Commodity;
+import fedorchuck.com.github.webstore.web.models.SearchRequest;
+import fedorchuck.com.github.webstore.dao.CommodityRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
