+++ src/main/java/fedorchuck/com/github/webstore/web/controllers/UserController.java
@@ -19,16 +19,16 @@
  * @license GPL-3.0+ <http://spdx.org/licenses/GPL-3.0+>
  */
 
-package fedorchuck.com.github.webstore.web;
+package fedorchuck.com.github.webstore.web.controllers;
 
 import static org.springframework.web.bind.annotation.RequestMethod.*;
 
 import javax.validation.Valid;
 
-import fedorchuck.com.github.webstore.AuthorizeUser;
-import fedorchuck.com.github.webstore.Commodity;
-import fedorchuck.com.github.webstore.SearchRequest;
-import fedorchuck.com.github.webstore.data.CommodityRepository;
+import fedorchuck.com.github.webstore.web.models.AuthorizeUser;
+import fedorchuck.com.github.webstore.domainmodels.Commodity;
+import fedorchuck.com.github.webstore.web.models.SearchRequest;
+import fedorchuck.com.github.webstore.dao.CommodityRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
@@ -36,8 +36,8 @@ import org.springframework.validation.Errors;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
 
-import fedorchuck.com.github.webstore.User;
-import fedorchuck.com.github.webstore.data.UserRepository;
+import fedorchuck.com.github.webstore.domainmodels.User;
+import fedorchuck.com.github.webstore.dao.UserRepository;
 import org.springframework.web.servlet.ModelAndView;
 
 import java.util.Collections;
