+++ src/main/java/fedorchuck/com/github/webstore/web/models/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * This can also be named {@code viewmodels}
+ */
+package fedorchuck.com.github.webstore.web.models;
