+++ src/main/resources/.gitignore
@@ -1,0 +1,1 @@
+/jdbc.properties
