+++ src/main/webapp/WEB-INF/views/index.html
@@ -1,6 +1,7 @@
 <html xmlns="http://www.w3.org/1999/xhtml"
       xmlns:th="http://www.thymeleaf.org">
 <head>
+        <meta charset="utf-8"/>
   <title>web store</title>
   <link rel="stylesheet"
         type="text/css"
