--- src/test/resources/jdbc.properties
@@ -1,6 +1,0 @@
-jdbc.driverClassName=org.postgresql.Driver
-jdbc.url=jdbc:postgresql://localhost:5432/webstore
-jdbc.username=postgres
-jdbc.password=lucky strike
-
-
